css/filter-effects/parsing/flood-opacity-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>Filter Effects Module Level 1: parsing flood-opacity with invalid values</title><metadata><h:link href="https://drafts.fxtf.org/filter-effects/#propdef-flood-opacity" rel="help"/><h:meta content="flood-opacity supports only the grammar 'alpha-value'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("flood-opacity", "1.");
test_invalid_value("flood-opacity", "2 3");

  ]]></script></svg>
